requirements.txt
opencv-python
numpy
pydantic
ultralytics
pyserial
pynput

